0.3.0.0
- Added Corner Data HUD(Alpha version/early access), showing corner delta and optionally: minimum speed, average speed and max lateral g force.
- DualSenseX module now allows to set a custom port for DSX(6969 is default).
- Lap table HUD: Reworked design and functionality, now shows personal fastest and best sectors based on valid laps.
- HUD Tab: Switched movement mode and always visible buttons around
- Revamped internals of several HUDs to lower cpu usage:
 - Lap Table HUD
 - Car Info HUD
 - Shared Memory page HUDs
- Wind Direction HUD: Added Wind Threshold slider for only showing the hud above a certain wind speed.

+++ Race_Element/Controls/Telemetry/RaceSessions/DivideBy1000ToFloatConverter.cs
@@ -10,7 +10,7 @@ namespace RaceElement.Controls.Telemetry.RaceSessions
         {
             if (value is int)
             {
-                float result = (int)value / 1000f;
+                double result = (int)value / 1000d;
                 return $"{result:F3}";
             }
             return string.Empty;
